LC-273: Renamed interface IBatch to Batch. Renamed implementation Batch to SingleBatch.
